backend/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
prometheus-fastapi-instrumentator>=6.1.0
